python_scripts/requirements.txt
dwave-ocean-sdk==6.1.1